sk_close
If we close() (or sigterm the application) in the middle of a send,
currently we don't send the rest of the bytes in the unencrypted or encrypted pages queues.

Normally tcp does do this if they are in the write buffer.

For the unencrypted pages (probably from MSG_MORE), override close, and
tls_push them.

For encrypted pages, sk_wmem_alloc won't go to zero until the tx_buff skb is freed, so write_space will be called, until skb destructor
sock_wfree is called from kfree_skb

This currently happens because we set_owner_w on the skb.  If we're moving to frags, we'd have to increment sk_wmem_alloc manually I think.